+++ rando/static/js/popup.js
@@ -5,12 +5,6 @@ $(document).ready(function (e) {
     // Links not leaving page
     $("#popup-home #popup-body a:not([href^='http'])").click(function () {
         $modal.modal('hide');
-        setTimeout(function () {
-            // Since a link can be a hash containing filters, and since
-            // we don't watch the hash location, we reload the filters
-            // after a click on such a link.
-            $(window).trigger('filters:reload');
-        }, 0);
     });
 
     // External links
@@ -24,10 +18,16 @@ $(document).ready(function (e) {
         $modal.modal('show');
     }
 
-    // Click on home shows popup
-    $('header a.home.popup').click(function () {
-        $modal.modal('show');
+    $(window).on('view:home view:detail', function () {
+        // Popup is shown only on homepage
+        $('header a.home.popup, header li.home a').click(function (e) {
+            if (/^\/[a-zA-Z_]{2,5}\/$/.test(window.location.pathname)) {
+                $modal.modal('show');
+                e.preventDefault();
+            }
+        });
     });
+
 
     function showModal() {
         var alreadyShown = localStorage.getItem('popup-shown') === "yes",