+++ rando/trekking/tests/test_popup.js
@@ -1,17 +1,11 @@
-casper.options.waitTimeout = 1000;
-casper.options.viewportSize = {width: 1280, height: 768};
-casper.userAgent("Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.11 (KHTML, like Gecko) Chrome/23.0.1271.97 Safari/537.11");
+var utils = require('./test_utils.js');
 
-casper.options.timeout = 60000;
-casper.options.onTimeout = function() {
-    casper.die("Timed out after 60 seconds.", 1);
-};
+utils.setUp();
 
 
 casper.test.begin('Welcome popup is only shown the first time', function(test) {
 
     var home_url = casper.cli.options['url-base'] + '/fr/';
-
 
     /*
      * Landing on home
@@ -19,71 +13,59 @@ casper.test.begin('Welcome popup is only shown the first time', function(test) {
 
     casper.start(home_url, function () {
         // On DOM ready.
-        casper.test.assertExists("header a.home.popup",
-                                 'Link to show popup is present');
+        test.assertExists("header a.home.popup",
+                          'Link to show popup is present');
     });
 
     casper.waitUntilVisible('#popup-home', function () {
+        test.pass('Popup is shown by default.');
         casper.click('#popup-home button.close');
     });
 
     casper.waitWhileVisible('#popup-home', function () {
-        casper.test.assertExists("#popup-home[aria-hidden='true']",
-                                 'Popup is now closed.');
+        test.assertExists("#popup-home[aria-hidden='true']",
+                          'Popup is now closed.');
     });
 
     casper.then(function () {
-        casper.log('Come back on home');
+        test.info('Come back on home');
         casper.open(home_url);
     });
 
-    casper.waitUntilVisible('#popup-home', undefined, function () {
-        casper.test.assertExists("#popup-home[aria-hidden='true']",
-                                 'Popup is not shown on second visit.');
+    casper.waitForSelector("#popup-home[aria-hidden='true']", function () {
+        test.pass('Popup is not shown on second visit.');
         casper.click('header a.home.popup');
-    }, 200);
-
-    casper.waitUntilVisible('#popup-home', function () {
-        casper.test.assertExists("#popup-home[aria-hidden='false']",
-                                 'Popup still opens on button click.');
     });
 
-    casper.then(clearLocalStorage);
+    casper.waitUntilVisible('#popup-home', function () {
+        test.assertExists("#popup-home[aria-hidden='false']",
+                          'Popup still opens on button click.');
+    });
+
+    casper.then(utils.clearLocalStorage);
 
     /*
      * Landing on other page than home
      */
     casper.then(function () {
-        casper.log('Open detail page');
+        test.info('Open detail page');
         casper.open(home_url + 'pages/flatpage');
     });
 
-    casper.waitUntilVisible('#popup-home', undefined, function () {
-        casper.test.assertExists("#popup-home[aria-hidden='true']",
-                                 'Popup is not shown if landing on other page.');
+    casper.waitForSelector("#popup-home[aria-hidden='true']", function () {
+        test.pass('Popup is not shown if landing on other page.');
+        casper.click('header a.home'); // Brings to home
+    });
 
-        casper.click('header a.home');
-    }, 200);
+    casper.waitForSelector('header li.home a', function () {
+        casper.click('header li.home a');
+    });
 
     casper.waitUntilVisible('#popup-home', function () {
-        casper.test.assertExists("#popup-home[aria-hidden='false']",
-                                 'Popup opens on button click.');
-        casper.test.assertUrlMatch(/fr\/#results$/, 'And home page is shown.');
+        test.assertExists("#popup-home[aria-hidden='false']",
+                          'Popup opens on button click.');
+        test.assertUrlMatch(home_url, 'And home page is shown.');
     });
 
-
-    casper.then(clearLocalStorage);  // For next sessions
-
-    casper.run(function () {
-        test.done();
-    });
-
-
-
-    function clearLocalStorage () {
-        // Clear localstorage for next sessions
-        casper.evaluate(function () {
-            localStorage.clear();
-        });
-    }
+    utils.done(test);
 });